commit15

+++ creator.c
@@ -17,7 +17,7 @@ size_t	add_int(int **new_mas, char **str, int **old_mas)
 	{
 		if (overflow_or_not(str[i]))
 			overflow_error(old_mas, new_mas, str);
-		(*new_mas)[len_new] = ft_atoi(str[i]);
+		(*new_mas)[len_new] = (int)ft_atoi(str[i]);
 		i++;
 		len_new++;
 	}
